code/backend/requirements.txt
# FastAPI框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# 缓存
redis==5.0.1
hiredis==2.2.3

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 工具
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2

# 对象存储
cos-python-sdk-v5==1.9.29

# 支付
wechatpy==1.8.18

# 其他
orjson==3.9.10
